device_discovery/requirements.txt
python_kasa==0.4.0.dev0
redis==3.5.3